# api/docs/swagger.yaml
basePath: /
definitions:
  controller.createUserReq:
    properties:
      email:
        type: string
      name:
        maxLength: 64
        minLength: 2
        type: string
    required:
    - email
    - name
    type: object
  controller.loginReq:
    properties:
      user_id:
        type: string
    required:
    - user_id
    type: object
  errcode.Code:
    properties:
      code:
        type: integer
      httpstatus:
        type: integer
      message:
        type: string
    type: object
  errcode.Error:
    properties:
      c:
        $ref: '#/definitions/errcode.Code'
      cause: {}
    type: object
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 基于 Go + Gin 的轻量级 Web 开发框架
  title: roc_way API
  version: "1.0"
paths:
  /api/v1/users:
    get:
      description: 获取所有用户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "500":
          description: 数据库错误
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: 用户列表
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: 创建新用户
      parameters:
      - description: 创建用户请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.createUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 数据库错误
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - Users
  /api/v1/users/{id}:
    get:
      description: 根据 ID 获取用户详情
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: 用户详情
      tags:
      - Users
  /auth/login:
    post:
      consumes:
      - application/json
      description: 根据 user_id 签发 JWT token
      parameters:
      - description: 登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: access_token & refresh_token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 用户登录
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 将 token 加入黑名单
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用 refresh_token 获取新的 access_token
      parameters:
      - description: 刷新请求
        in: body
        name: request
        required: true
        schema:
          properties:
            refresh:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 新的 access_token & refresh_token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "401":
          description: Token无效
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 刷新 Token
      tags:
      - Auth
  /healthz:
    get:
      description: 返回服务健康状态
      produces:
      - application/json
      responses:
        "200":
          description: status":"ok
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 健康检查
      tags:
      - Health
  /sse/notifications:
    get:
      description: 通过 Server-Sent Events 推送实时通知
      produces:
      - text/event-stream
      responses:
        "200":
          description: event-stream
          schema:
            type: string
      summary: SSE 实时通知
      tags:
      - Realtime
  /ws/chat:
    get:
      description: 通过 WebSocket 进行实时双向通信
      responses:
        "101":
          description: ' Switching Protocols'
          schema:
            type: string
      summary: WebSocket 聊天
      tags:
      - Realtime
swagger: "2.0"